+++ docs/whats_new.rst
@@ -15,6 +15,7 @@ v0.7.0 (unreleased)
 
 Breaking Changes
 ~~~~~~~~~~~~~~~~
+- Removed support for Python 2. This is the first version of regionmask that is Python 3 only!
 
 - ``matpoltlib`` and ``cartopy`` are now optional dependencies (:issue:`169`).
 
@@ -35,6 +36,8 @@ Enhancements
 - :py:attr:`Regions.plot()` and :py:attr:`Regions.plot_regions()` now take the
   ``label_multipolygon`` keyword to add text labels to all Polygons of
   MultiPolygons (:issue:`185`).
+- :py:attr:`Regions.plot()` and :py:attr:`Regions.plot_regions()` now warn on unused arguments,
+  e.g. ``plot(add_land=False, land_kws=dict(color="g"))`` (:issue:`192`).
 
 New regions
 ~~~~~~~~~~~